Conflicts:
	WebContent/Login.jsp
	src/controller/Login.java
	src/controller/Registry.java

+++ src/model/service/ArticleClassService.java
@@ -3,48 +3,35 @@ package model.service;
 import model.dao.ArticleClassDAO;
 import model.dao.jdbc.ArticleClassDAOjdbc;
 import model.vo.ArticleClassVO;
+
 /**
  * @author iTV小組成員
  *
  */
 public class ArticleClassService {
 	private ArticleClassDAO dao;
+
 	/**
 	 * 初始化ArticleClassDAO
 	 */
 	public ArticleClassService() {
 		this.dao = new ArticleClassDAOjdbc();
 	}
-	
-	/**
-	 * 增加文章類別
-	 * 
-	 * @param subclassNo 文章子分類編號
-	 * @param subclassName 文章子分類名稱
-	 * @param className 文章類別名稱
-	 * @return true 增加成功; false 增加失敗
-	 */	
-	public boolean addArticleClass(String subclassNo, String subclassName, String className) {
-		ArticleClassVO bean = new ArticleClassVO();
-		bean.setSubclassNo(subclassNo);
-		bean.setSubclassName(subclassName);
-		bean.setClassName(className);
-		return dao.insert(bean);
-	}
 
 	/**
 	 * 增加文章類別
 	 * 
-	 * @param bean 必須包含 <b>subclassNo</b>、 <b>subclassName</b> 與 <b>className</b>
+	 * @param bean
+	 *            必須包含 <b>subclassNo</b>、 <b>subclassName</b> 與 <b>className</b>
 	 * @return true 增加成功; false 增加失敗
-	 */	
+	 */
 	public boolean addArticleClass(ArticleClassVO bean) {
 		boolean result = false;
-		bean.getSubclassNo();
-		bean.getSubclassName();
-		bean.getClassName();
 		if (bean != null) {
-			result = dao.insert(bean);
+			int temp = dao.insert(bean);
+			if (temp == 1) {
+				result = true;
+			}
 		}
 		return result;
 	}
@@ -52,32 +39,17 @@ public class ArticleClassService {
 	/**
 	 * 修改文章類別
 	 * 
-	 * @param subclassNo 文章子分類編號
-	 * @param subclassName 文章子分類名稱
-	 * @param className 文章類別名稱
+	 * @param bean
+	 *            必須包含 <b>SubclassNo</b>、<b>SubclassNo</b>、<b>SubclassNo</b>
 	 * @return true 修改成功; false 修改失敗
-	 */		
-	public boolean updateArticleClass(String subclassNo, String subclassName, String className) {
-		ArticleClassVO bean = new ArticleClassVO();
-		bean.setSubclassNo(subclassNo);
-		bean.setSubclassName(subclassName);
-		bean.setClassName(className);
-		return dao.update(bean);
-	}
-
-	/**
-	 * 修改文章類別
-	 * 
-	 * @param bean 必須包含 <b>SubclassNo</b>、<b>SubclassNo</b>、<b>SubclassNo</b>
-	 * @return true 修改成功; false 修改失敗
-	 */		
+	 */
 	public boolean updateArticleClass(ArticleClassVO bean) {
 		boolean result = false;
-		bean.getSubclassNo();
-		bean.getSubclassName();
-		bean.getClassName();
 		if (bean != null) {
-			result = dao.update(bean);
+			int temp = dao.update(bean);
+			if (temp == 1) {
+				result = true;
+			}
 		}
 		return result;
 	}
@@ -85,10 +57,18 @@ public class ArticleClassService {
 	/**
 	 * 刪除文章類別
 	 * 
-	 * @param subclassNo 文章子類別編號
+	 * @param subclassNo
+	 *            文章子類別編號
 	 * @return true 刪除成功; false 刪除失敗
-	 */		
+	 */
 	public boolean deleteArticleClass(String subclassNo) {
-		return dao.delete(subclassNo);
+		boolean result = false;
+		if (subclassNo != null && subclassNo.trim().length() != 0) {
+			int temp = dao.delete(subclassNo);
+			if (temp == 1) {
+				result = true;
+			}
+		}
+		return result;
 	}
 }
